sentiment_analysis_dashboard/backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Machine Learning
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.24.4
pandas==2.1.4

# Data validation
pydantic==2.5.0
email-validator==2.1.0

# System monitoring
psutil==5.9.6

# Environment and configuration
python-dotenv==1.0.0

# HTTP client (for testing)
httpx==0.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0